requirements.txt
bcrypt==5.0.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
docker==7.1.0
idna==3.11
invoke==2.2.1
paramiko==4.0.0
pycparser==2.23
PyNaCl==1.6.1
python-dotenv==1.2.1
requests==2.32.5
tabulate==0.9.0
urllib3==2.5.0
